+++ dao/public.js
@@ -20,7 +20,7 @@ const err = require(resolve(__dirname, "..", "errors"));
  * @param  {Object{_id, upload}}    _document [用户信息]
  * @return {Generator}           [数据库操作函数，使用co或next来驱动]
  */
-function update (_document) {
+function* update (_document) {
   const {_id, upload} = _document;
   if (this.valid(_id)) {
     return this.updateOne({_id}, upload);
@@ -34,7 +34,7 @@ function update (_document) {
  * @param  {Object{name}}    _user [_document]
  * @return {Generator}           [数据库操作函数，使用co或next来驱动]
  */
-function insert (_document) {
+function* insert (_document) {
   const {name} = _document;
   if (!name) {
     err.throwLackParameters();
@@ -58,9 +58,9 @@ function insert (_document) {
  * @param  {Array}    _documents  [需要插入的一群数据]
  * @return {Generator}            [数据库操作函数，使用co或next来驱动]
  */
-function insertMany (_documents) {
+function* insertMany (_documents) {
   if (_documents instanceof Array) {
-    return this.insertMany(_documents);
+    return _documents.length > 0 ? this.insertMany(_documents): null;
   }
   err.throwParametersError();
 }
@@ -71,7 +71,7 @@ function insertMany (_documents) {
  * @param  {Object{_id}}    _document [需要删除的数据]
  * @return {Generator}                [数据库操作函数，使用co或next来驱动]
  */
-function remove ({_id}) {
+function* remove ({_id}) {
   if (this.valid(_id)) {
     return this.updateOne({_id}, {
       $set: {
@@ -88,16 +88,9 @@ function remove ({_id}) {
  * @param  {Object{name}} _document   [需要检测的数据]
  * @return {[type]}           [description]
  */
-function checkExist ({name}){
-  const __self = this;
-  return new Promise((solve, reject) => {
-    __self.findOne({name}).then(_doc => {
-      if (_doc) {
-        solve(true);
-      }
-      solve(false);
-    }).catch(err => reject(err));
-  });
+function* checkExist ({name}){
+  const _doc = yield this.findOne({name});
+  return _doc ? true : false;
 }
 
 module.exports = {