+++ test/userService_test.js
@@ -11,10 +11,15 @@
   */
 "use strict";
 
-const {resolve} = require("path");
+const {resolve} = require("path"), co = require("co");
+const {user_dao} = require(resolve(__dirname, "..", "dao"));
 const userService = require(resolve(__dirname, "..", "services", "userService"));
 
 const user_xlsx = resolve(process.env.HOME, "user.xlsx");
 const users = userService.loaderFromXlsx(user_xlsx);
 
-console.log(users);
+co(function* () {
+  const p = yield userService.createUser(users);
+  console.log(p);
+  user_dao.close();
+});